Join device services. Minor fixes.
