--- Sources/Indice.Identity/Client/Services/DeviceRegistrationService.swift
@@ -1,370 +1,0 @@
-//
-//  IdentityClientDeviceRegistration.swift
-//  Indice.Identity
-//
-//  Created by Nikolas Konstantakopoulos on 30/3/23.
-//
-
-import Foundation
-import Combine
-
-
-fileprivate extension ValueStorageKey {
-    static var devicePinKey   = ValueStorageKey(name: "device_registration_device_pin" )
-    static var fingerprintKey = ValueStorageKey(name: "device_registration_fingerprint")
-}
-
-
-
-public class DeviceStatus: ObservableObject {
-    @Published
-    public internal(set)
-    var hasDevicePin: Bool = false
-    
-    @Published
-    public internal(set)
-    var hasFingerprint: Bool = false
-    
-    @Published
-    public internal(set)
-    var hasQuickLogin: Bool = false
-
-}
-
-/** 
- Service responsible for managing the current device's authorization state.
- Registers and removes anything necessary for fingerprint and 4pin authorization.
- */
-public protocol DeviceRegistrationService {
-    
-    var deviceStatus: DeviceStatus { get }
-    
-    /** Register or update a device, to be able to perform a **device\_authentication** grant with **pin** mode */
-    @available(*, deprecated, message: "Use the  registerDevice(withPin:otpChannel:) async throws -> (CallbackType.OtpResult) async throws -> ()")
-    func registerDevice(withPin pin: String, otpChannel: TotpDeliveryChannel?, otpProvider: CallbackType.OtpProvider) async throws
-        
-    /**
-     Register or update a device, to be able to perform a **device\_authentication** grant with **pin** mode
-     This method returns a **lambda** that, provided with on otp, completes the device pin registration.
-     ```
-     
-     let pin: String = makePin()
-     let continuation = try await registerDevice(withPin: pin)
-     
-     // If success show otp input
-     try await continuation(otpResult)
-     
-     ```
-     It is not necessary to call the continuation with an OtpResult.aborted result but it is recommended.
-     */
-    func registerDevice(withPin pin: String, otpChannel: TotpDeliveryChannel?) async throws -> (CallbackType.OtpResult) async throws -> ()
-    
-    
-    @available(*, deprecated, message: "Use the registerDeviceFingerprint(otpChannel:) async throws -> (CallbackType.OtpResult) async throws -> ()")
-    /** Register or update a device, to be able to perform a **device\_authentication** grant with **fingerprint** mode */
-    func registerDeviceFingerprint(otpChannel: TotpDeliveryChannel?, otpProvider: CallbackType.OtpProvider) async throws
-    
-    /**
-     Register or update a device, to be able to perform a **device\_authentication** grant with **fingerprint** mode
-     This method returns a **lambda** that, provided with on otp, completes the device fingerprint registration.
-     ```
-     
-     let continuation = try await registerDeviceFingerprint()
-     // If success show otp input
-     try await continuation(otpResult)
-     
-     ```
-     It is not necessary to call the continuation with an OtpResult.aborted result but it is recommended.
-     */
-    func registerDeviceFingerprint(otpChannel: TotpDeliveryChannel?) async throws -> (CallbackType.OtpResult) async throws -> ()
-    
-    /** Remove a device pin registration */
-    func removeRegistrationDevicePin() async
-    
-    /** Remove a fingerprint registration */
-    func removeRegistrationFingerprint() async
-    
-    /** Trigger enable current device's trust status */
-    func enableDeviceTrust(deviceSelection: CallbackType.DeviceSelection) async throws
-    
-    /** Remove current device's trust status */
-    func removeDeviceTrust() async throws
-}
-
-
-internal class DeviceRegistrationServiceImpl: DeviceRegistrationService {
-
-    private let thisDeviceRepository : ThisDeviceRepository
-    private let devicesRepository    : DevicesRepository
-    private let valueStorage         : ValueStorage
-    private let client               : Client
-    private let devicesService       : DevicesServiceInternal
-    private var cancellables         = Set<AnyCancellable>()
-    
-    public
-    private(set)
-    var deviceStatus: DeviceStatus = .init()
-    
-    init(valueStorage: ValueStorage, 
-         client: Client,
-         devicesRepository: DevicesRepository,
-         thisDeviceRepository: ThisDeviceRepository,
-         devicesService: DevicesServiceInternal) {
-        self.valueStorage = valueStorage
-        self.client = client
-        self.devicesRepository = devicesRepository
-        self.thisDeviceRepository = thisDeviceRepository
-        self.devicesService = devicesService
-        
-        self.deviceStatus.hasDevicePin   = valueStorage.readBool(forKey: .devicePinKey)   ?? false
-        self.deviceStatus.hasFingerprint = valueStorage.readBool(forKey: .fingerprintKey) ?? false
-        
-        self.deviceStatus
-            .$hasFingerprint
-            .sink { [weak self] newValue in
-                self?.valueStorage.store(value: newValue, forKey: .fingerprintKey)
-            }.store(in: &cancellables)
-        
-        self.deviceStatus.$hasDevicePin
-            .sink { [weak self] newValue in
-                self?.valueStorage.store(value: newValue, forKey: .devicePinKey)
-            }.store(in: &cancellables)
-        
-        self.deviceStatus.$hasFingerprint
-            .combineLatest(self.deviceStatus.$hasDevicePin)
-            .map { $0 || $1 }
-            .sink(receiveValue: { [weak self] result in
-                self?.deviceStatus.hasQuickLogin = result
-            })
-            .store(in: &cancellables)
-    }
-    
-    public func registerDevice(withPin pin: String, otpChannel: TotpDeliveryChannel? = nil, otpProvider: CallbackType.OtpProvider) async throws {
-        do {
-            _ = CryptoUtils.deleteKeyPair(locked: false, tagged: .devicePin)
-            let keys = try CryptoUtils.createKeyPair(locked: false, tagged: .devicePin)
-            
-            let verifier     = CryptoRandom.uniqueId()
-            let verifierHash = CryptoUtils.challenge(for: verifier)
-            let deviceIds    = thisDeviceRepository.ids
-            let deviceInfo   = thisDeviceRepository.info
-            let devicePin    = try CryptoUtils.prepare(pin: pin, withDeviceId: deviceIds.device, and: keys)
-            /* TODO: This request could have an OTP side-effect.
-             Find a nice way to decide if it will and pass it to the OTP provider?
-             */
-            let response = try await devicesRepository.initialize(authRequest: .pinInit(codeChallenge: verifierHash,
-                                                                                       deviceIds: deviceIds,
-                                                                                       client: client))
-            
-            let signedVerifier = try CryptoUtils.sign(string: response.challenge, with: keys)
-            
-            // TODO: pass the status of the otp need.
-            let otpResult = await otpProvider(true)
-            
-            
-            // TODO: Handle errors and stuff
-            // Should this throw something?
-            if otpResult.isAborted { deviceStatus.hasDevicePin = false; return }
-            
-            let registration = try await devicesRepository.complete(registrationRequest: .pin(code: response.challenge,
-                                                                                             codeVerifier: verifier,
-                                                                                             codeSignature: signedVerifier,
-                                                                                             deviceIds: deviceIds,
-                                                                                             deviceInfo: deviceInfo,
-                                                                                             devicePin: devicePin,
-                                                                                             otp: otpResult.otpValue))
-            
-            thisDeviceRepository.update(registrationId: registration.registrationId)
-        } catch {
-            deviceStatus.hasDevicePin = false
-            throw error
-        }
-        
-        deviceStatus.hasDevicePin = true
-    }
-    
-    
-    public func registerDevice(withPin pin: String, otpChannel: TotpDeliveryChannel?) async throws -> (CallbackType.OtpResult) async throws -> () {
-        do {
-            _ = CryptoUtils.deleteKeyPair(locked: false, tagged: .devicePin)
-            let keys = try CryptoUtils.createKeyPair(locked: false, tagged: .devicePin)
-            
-            let verifier     = CryptoRandom.uniqueId()
-            let verifierHash = CryptoUtils.challenge(for: verifier)
-            let deviceIds    = thisDeviceRepository.ids
-            let deviceInfo   = thisDeviceRepository.info
-            let devicePin    = try CryptoUtils.prepare(pin: pin, withDeviceId: deviceIds.device, and: keys)
-            /* TODO: This request could have an OTP side-effect.
-             Find a nice way to decide if it will and pass it to the OTP provider?
-             */
-            let response = try await devicesRepository.initialize(authRequest: .pinInit(codeChallenge: verifierHash,
-                                                                                       deviceIds: deviceIds,
-                                                                                       client: client))
-            
-            let signedVerifier = try CryptoUtils.sign(string: response.challenge, with: keys)
-            
-            return { [weak self] otpResult in
-                guard let self = self else { return }
-                do {
-                    let registration = try await devicesRepository.complete(registrationRequest: .pin(code: response.challenge,
-                                                                                                     codeVerifier: verifier,
-                                                                                                     codeSignature: signedVerifier,
-                                                                                                     deviceIds: deviceIds,
-                                                                                                     deviceInfo: deviceInfo,
-                                                                                                     devicePin: devicePin,
-                                                                                                     otp: otpResult.otpValue))
-                    
-                    self.thisDeviceRepository.update(registrationId: registration.registrationId)
-                } catch {
-                    self.deviceStatus.hasFingerprint = false
-                    throw error
-                }
-            }
-        } catch {
-            deviceStatus.hasFingerprint = false
-            throw error
-        }
-    }
-    
-    
-    public func registerDeviceFingerprint(otpChannel: TotpDeliveryChannel? = nil) async throws -> (CallbackType.OtpResult) async throws -> () {
-        do {
-            _ = CryptoUtils.deleteKeyPair(locked: true, tagged: .fingerprint)
-            let keys = try CryptoUtils.createKeyPair(locked: true, tagged: .fingerprint)
-            
-            let verifier     = CryptoRandom.uniqueId()
-            let verifierHash = CryptoUtils.challenge(for: verifier)
-            let deviceIds    = thisDeviceRepository.ids
-            let deviceInfo   = thisDeviceRepository.info
-            let devicePem    = try CryptoUtils.pem(for: keys)
-            
-            /* TODO: This request could have an OTP side-effect.
-             Find a nice way to decide if it will and pass it to the OTP provider?
-             */
-            
-            let response = try await devicesRepository.initialize(authRequest: .biometrictInit(codeChallenge: verifierHash,
-                                                                                              deviceIds: deviceIds,
-                                                                                              client: client))
-            let signedVerifier = try CryptoUtils.sign(string: response.challenge, with: keys)
-
-            return { [weak self] otpResult in
-                guard let self = self else { return }
-                do {
-                    let registration = try await self.devicesRepository.complete(registrationRequest: .biometric(code: response.challenge,
-                                                                                                                codeVerifier: verifier,
-                                                                                                                codeSignature: signedVerifier,
-                                                                                                                deviceIds: deviceIds,
-                                                                                                                deviceInfo: deviceInfo,
-                                                                                                                publicPem: devicePem,
-                                                                                                                otp: otpResult.otpValue))
-                    
-                    self.thisDeviceRepository.update(registrationId: registration.registrationId)
-                } catch {
-                    self.deviceStatus.hasFingerprint = false
-                    throw error
-                }
-            }
-        } catch {
-            deviceStatus.hasFingerprint = false
-            throw error
-        }
-    }
-    
-
-    public func registerDeviceFingerprint(otpChannel: TotpDeliveryChannel? = nil, otpProvider: CallbackType.OtpProvider) async throws {
-        do {
-            _ = CryptoUtils.deleteKeyPair(locked: true, tagged: .fingerprint)
-            let keys = try CryptoUtils.createKeyPair(locked: true, tagged: .fingerprint)
-            
-            let verifier     = CryptoRandom.uniqueId()
-            let verifierHash = CryptoUtils.challenge(for: verifier)
-            let deviceIds    = thisDeviceRepository.ids
-            let deviceInfo   = thisDeviceRepository.info
-            let devicePem    = try CryptoUtils.pem(for: keys)
-            
-            /* TODO: This request could have an OTP side-effect.
-             Find a nice way to decide if it will and pass it to the OTP provider?
-             */
-            
-            let response = try await devicesRepository.initialize(authRequest: .biometrictInit(codeChallenge: verifierHash,
-                                                                                               deviceIds: deviceIds,
-                                                                                               client: client))
-            let signedVerifier = try CryptoUtils.sign(string: response.challenge, with: keys)
-            
-            // TODO: pass the status of the otp need.
-            let otpResult = await otpProvider(true)
-            
-            
-            // TODO: Should this throw something?
-            if otpResult.isAborted { deviceStatus.hasFingerprint = false; return }
-            
-            let registration = try await devicesRepository.complete(registrationRequest: .biometric(code: response.challenge,
-                                                                                                    codeVerifier: verifier,
-                                                                                                    codeSignature: signedVerifier,
-                                                                                                    deviceIds: deviceIds,
-                                                                                                    deviceInfo: deviceInfo,
-                                                                                                    publicPem: devicePem,
-                                                                                                    otp: otpResult.otpValue))
-            
-            thisDeviceRepository.update(registrationId: registration.registrationId)
-        } catch {
-            deviceStatus.hasFingerprint = false
-            throw error
-        }
-        deviceStatus.hasFingerprint = true
-    }
-    
-    public func removeRegistrationDevicePin() async {
-        CryptoUtils.deleteKeyPair(locked: false, tagged: .devicePin)
-        deviceStatus.hasDevicePin = false
-    }
- 
-    public func removeRegistrationFingerprint() async {
-        CryptoUtils.deleteKeyPair(locked: true, tagged: .fingerprint)
-        deviceStatus.hasFingerprint = false
-    }
- 
-    public func enableDeviceTrust(deviceSelection: CallbackType.DeviceSelection) async throws {
-        let ids = thisDeviceRepository.ids
-        
-        if devicesService.devicesInfo.userDevices == nil {
-            try await devicesService.refreshDevices()
-        }
-        
-        let maxTrustedCount = 1 // TODO: Get from API?
-        let devices = (devicesService.devicesInfo.userDevices ?? []).filter {
-            $0.deviceId != ids.device
-        }
-            
-        let currentTrustedCount = devices.filter {
-            $0.isTrusted == true
-        }.count
-        
-        let swapDeviceId: String? = await {
-            if currentTrustedCount >= maxTrustedCount {
-                if case .swap(let deviceInfo) = await deviceSelection(devices) {
-                    return deviceInfo.deviceId
-                } else {
-                    // TODO: this should throw something here?
-                    return nil
-                }
-            }
-            
-            return nil
-        }()
-        
-        try await devicesRepository.trust(deviceId: ids.device, bySwappingWith: swapDeviceId)
-        
-        try await devicesService.updateDeviceWith(deviceId: ids.device)
-        
-        if let swapDeviceId { try await devicesService.updateDeviceWith(deviceId: swapDeviceId) }
-    }
-    
-    public func removeDeviceTrust() async throws {
-        let deviceId = thisDeviceRepository.ids.device
-        
-        try await devicesRepository.unTrust(deviceId: deviceId)
-        try await devicesService.updateDeviceWith(deviceId: deviceId)
-    }
-    
-}
